MINOR: Use HAProxy's core.concat function
Lua concatenation function 'str' .. 'str' is pretty ineficient when we
have to append data many times data to a string in a loop. This is the
case with the buffer variable: it is filled up with the metrics
requested by the client.

This patch make the built of the buffer more efficient by using
core.concat() function.