MAJOR: implementing output filtering
By default, this Lua prometheus exporter exports all available
statistics for every frontends and backends.

This patch allows filtering the ouptput based on query string
parameters.
Now, we can say "I want metrics for this backend and this backend" or "I
want metrics for this backend and this frontend" or "I want this metric
and this metric for this backend" or "I want this metric only"
Check README for URL examples.