MAJOR: Add a 'transform' function to HAProxy metrics
The purpose of the 'transform' function is to convert an HAProxy metric
value into something prometheus can consume.
Not all metric needs it and first implementation in the code was doing
string matching on a per case basis. This is quite ineficient.

From now, if a field needs to be translated, simply create the function
which takes as an input a value, returns the transformed value, and add
a table pointer to that function into the right metric table field.